LPD-86378 Update liferay/js-api use. Merge both getFDSAtom functionalities. Integrate State type definition into current file rather than creating a new one
