+++ modules/apps/frontend-data-set/frontend-data-set-sample-web/src/main/resources/META-INF/resources/js/fragments/AdvancedSearch.tsx
@@ -6,7 +6,7 @@
 import ClayButton from '@clayui/button';
 import {ClayInput} from '@clayui/form';
 import ClayLayout from '@clayui/layout';
-import {IFDSState, getFDSAtom} from '@liferay/frontend-data-set-web';
+import {IFDSState, getOrCreateFDSAtom} from '@liferay/frontend-data-set-web';
 import {useLiferayState} from '@liferay/frontend-js-state-web/react';
 import React, {useEffect, useMemo, useState} from 'react';
 
@@ -22,7 +22,7 @@ const ADVANCED_FDS_ATOM_KEY = 'advancedFDSAtom';
  */
 const AdvancedSearch = () => {
 	const memoizedAtom = useMemo(
-		() => getFDSAtom({atomKey: ADVANCED_FDS_ATOM_KEY}),
+		() => getOrCreateFDSAtom({atomKey: ADVANCED_FDS_ATOM_KEY}),
 		[]
 	);
