+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/java/com/liferay/frontend/data/set/web/internal/frontend/js/importmaps/extender/FDSConnectionDynamicJSImportMapsContributor.java
@@ -1,5 +1,5 @@
 /**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
