+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/FrontendDataSet.tsx
@@ -60,7 +60,7 @@ import {readConfigFromURL} from './utils/configInURL';
 import EVENTS from './utils/eventsDefinitions';
 import {activateFilter} from './utils/filters/activateFilter';
 import {deactivateFilter} from './utils/filters/deactivateFilter';
-import {getFDSAtom} from './utils/getFDSAtom';
+import {getOrCreateFDSAtom} from './utils/getFDSAtom';
 import getRandomId from './utils/getRandomId';
 
 // @ts-ignore
@@ -368,7 +368,7 @@ const FrontendDataSetContent = ({
 	});
 
 	const memoizedAtom = useMemo(
-		() => getFDSAtom({atom, fdsName: id}),
+		() => getOrCreateFDSAtom({atom, fdsName: id}),
 		[atom, id]
 	);
