+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/utils/getFDSAtom.ts
@@ -1,5 +1,5 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
@@ -7,7 +7,79 @@ import {Atom, Selector, State} from '@liferay/frontend-js-state-web';
 
 import {IFDSState} from './types';
 
-const getFDSAtom = ({
+import type {FDSState} from '@liferay/js-api/data-set';
+
+const DEFAULT_TIMEOUT = 5000;
+const DEFAULT_INTERVAL = 100;
+
+/**
+ * Waits for the FDS atom that the widget registers under
+ * `${id}_fdsState` to appear in the global `Liferay.State` registry,
+ * polling until it is found or the timeout elapses.
+ *
+ * The runtime must go through the global singleton (not the imported
+ * copy of `@liferay/frontend-js-state-web`) so that consumers share the
+ * same atom registry that the Frontend Data Set widget populates.
+ */
+export function getFDSAtom(
+	id: string,
+	options?: {interval?: number; timeout?: number}
+): Promise<Liferay.State.Atom<FDSState>> {
+	const timeout = options?.timeout ?? DEFAULT_TIMEOUT;
+	const interval = options?.interval ?? DEFAULT_INTERVAL;
+
+	const key = `${id}_fdsState`;
+
+	return new Promise((resolve, reject) => {
+		const existing = Liferay?.State?.__unsafe__?.getAtomOrSelectorKey(key);
+
+		if (existing) {
+			return resolve(existing as Liferay.State.Atom<FDSState>);
+		}
+
+		const startTime = Date.now();
+
+		const poll = setInterval(() => {
+			const atom = Liferay?.State?.__unsafe__?.getAtomOrSelectorKey(key);
+
+			if (atom) {
+				clearInterval(poll);
+
+				return resolve(atom as Liferay.State.Atom<FDSState>);
+			}
+
+			if (Date.now() - startTime >= timeout) {
+				clearInterval(poll);
+
+				reject(
+					new Error(
+						`FDS atom "${key}" was not found within ${timeout}ms`
+					)
+				);
+			}
+		}, interval);
+	});
+}
+
+export function getOrCreateSelector<T>(
+	key: string,
+	deriveValue: (get: Liferay.State.Getter) => T
+): Liferay.State.Selector<T> {
+	const existing = Liferay.State.__unsafe__.getAtomOrSelectorKey(key);
+
+	return (
+		(existing as Liferay.State.Selector<T> | null) ??
+		Liferay.State.selector<T>(key, deriveValue)
+	);
+}
+
+/**
+ * Creates, or returns the already-registered, FDS atom used by the
+ * widget. Unlike `getFDSAtom`, this is synchronous and is the producer
+ * of the atom: it registers a new atom under `${fdsName}_fdsState` when
+ * none exists yet.
+ */
+export function getOrCreateFDSAtom({
 	atom,
 	atomKey,
 	fdsName,
@@ -15,10 +87,10 @@ const getFDSAtom = ({
 	atom?: Atom<IFDSState> | undefined;
 	atomKey?: string;
 	fdsName?: string;
-}): Atom<IFDSState> | Selector<IFDSState> => {
+}): Atom<IFDSState> | Selector<IFDSState> {
 	if (!atom && !atomKey && !fdsName) {
 		throw new Error(
-			'getFDSAtom requires at least one of the following parameters: atom, atomKey, fdsName'
+			'getOrCreateFDSAtom requires at least one of the following parameters: atom, atomKey, fdsName'
 		);
 	}
 
@@ -40,6 +112,4 @@ const getFDSAtom = ({
 			search: {query: ''},
 		})
 	);
-};
-
-export {getFDSAtom};
+}
