+++ modules/apps/frontend-js/frontend-js-web/src/main/resources/META-INF/resources/liferay/liferay.d.ts
@@ -174,6 +174,24 @@ declare module Liferay {
 	}
 
 	namespace State {
+		interface Atom<T> {
+			readonly __type?: T;
+			readonly key: string;
+		}
+
+		interface Selector<T> {
+			readonly __type?: T;
+			readonly key: string;
+		}
+
+		type Getter = <T>(atomOrSelector: Atom<T> | Selector<T>) => T;
+
+		namespace __unsafe__ {
+			function getAtomOrSelectorKey(
+				key: string
+			): Atom<unknown> | Selector<unknown> | null;
+		}
+
 		type Primitive =
 			| bigint
 			| boolean
