+++ workspaces/liferay-sample-workspace/client-extensions/liferay-sample-custom-element-7/tsconfig.json
@@ -11,7 +11,7 @@
 		"noEmit": true,
 		"paths": {
 			"@liferay/frontend-data-set-web/api": [
-				"./node_modules/@liferay/js-api/data-set/client"
+				"./node_modules/@liferay/js-api/data-set/connection"
 			]
 		},
 		"skipLibCheck": true,
